Basic working functionality

+++ Server/NetworkWrapper/NetPeerWrapper.cs
@@ -30,7 +30,7 @@ namespace GameServer
 
     public void SendPayload(byte[] buffer, int length)
     {
-      this.peer.SendPayload(buffer, length);
+      this.peer.SendPayload(buffer, (ushort)length);
     }
   }
 }
